[JS] extracted keyrings.js and extrinsics.js
